Func: add initial_log_id and new_log_id
Log id is used to stand for a process such as a write req process or a
recovery operation. The log id can be used to track such process.

Log id will be initialized when a deamon(osd, mon, client) started by a
prefix and a random num, the g_log_id will be set to prefix << 10 +
random.

Each process can use new_log_id() function to get a new log id.

Signed-off-by: Dong Yuan <yuandong1222@gmail.com>